Change to nwfsc-assess/nwfscSurvey from JRWToolBox
Change to downloading the data using Chantel's survey
package versus JRWToolBox because I have access to the former
but not the later with respect to editing the code.
The functions return the same data except that the former is
subset for depth the same way that Beth provided data.
It could be good to think about this in terms of the other surveys.
Not sure if we would want deeper depths for the slope surveys.