Add compile to global search @seananderson pkgload helped
@seananderson provided a pull request with commit
265ef7d5eb59184860385828f5374fe0e246ef63
that provided information on how to add compile to the global env
using assign.
The package devtools uses pkgload that has a specific function version
of system.file that they add to the search path using an environment
and the 'shim_' notation. I used a combination of @seananderson
suggestion of searching and then the pkgload use of an environment.
The add is in VAST_run, which I believe is the only function that I have
that actually calls compile. This will only be an issue for the first time
that you compile, after that the library is loaded and you are fine.
This fix was needed because VAST doesn't specify which compile to use
because VAST allows users to load either TMB or TMBdebug to specify which
version of compile that you want to use. I wanted to stop users from having
to specifically load TMB using library(TMB) or library(TMBdebug) prior to
running any function within VASTWestCoast. I tested the minimal example
and it still works. See the pkgload for reference
https://github.com/r-lib/pkgload/blob/0ef4f58cd8135e9e77c21243e466b9739e608bde/R/shims.r
and
https://github.com/r-lib/pkgload/blob/0ef4f58cd8135e9e77c21243e466b9739e608bde/R/load.r
Was able to remove library(TMB) from the tests.