Update WCGOP example 2009 or newer and only longnose
Vlada talked to those who use the data and they recommended that only
data for longnose since 2009 should be used because they were sorted
to species level after the 2007 assessment.
Still need to find a way to sort out some of the zero observations
because the dataset is too big for VAST without the second more
arbitrary subsetting. Vlada will be working on this.