Implement 3 settings field, rho, and comp
field allows for configurations of FieldCongif
rho allows for configurations of RhoConfig
and comp is a logical that will allow for estimating comp data
Comp is not currently used, but I will be adding features later.
Field and rho take either a single value or a vector of four values
and use that to set the VAST parameters. Rho was needed for the Triennial
and Field was needed for the WCGOP data.