Add clean_data to separate adding columns from VAST_condition
The new function combines cleaning the data from the alaska center
and from the nw center into a single function based on the column
names that are present. This should work on an external data set that
is downloaded when someone wants two or more species but wants to treat
those species as a single species, i.e., Vlada's request. It also
removed the dependency on ThorsonUtilities, which is why VAST_EM
was included in this commit.