Change to onLoad rather than onAttach
Not sure if I even need INLA in here now b/c I
made it a dependent package. It doesn't hurt though.
Also no longer need James-Thorson packages b/c
I have remotes for them. The issue was with INLA not
bringing up the window to set a mirror when it was
being installed through James-Thorson/package, but he
added similar code to what I have here to fix it for me in
his development branches.
The change to onLoad is so that it will be called when someone
uses a function for the first time rather than just when
library(VASTWestCoast) gets called. This seemed more ideal
because I use the function calls all the time for a package
rather than loading the entire thing.