refactor: clean_data for Triennial instead of in VAST_condition
Vessel code was a little messy. This commit cleans it up and
assigns Vessel to none then adds year for the Triennial. The
raw column is no longer created b/c clean_data gives the raw
column a new name anyway.

Removes 1977 because of inconsistent sampling methods. This was
removed before, but now it is in clean_data.