build: clean up VAST_setup and VAST_condition
Remove some commented code
Remove an object that doesn't need to be assigned
Use get_extraplist that removes list from setup to standalone code
Change region to survey for consistency
Remove extrapolation argument in get_settings because
it should be consistent for surveys. Triennial is based on
propInTriennial.
Move repetative code to VAST_do, which is not exported
This is a good way to run the sensitivities for the Triennial.
VAST_condition had some bad documentation
Save DatabaseSave in VAST_do so diagnostics work
Update some more documentation in NAMESPACE remove not needed package