Fix: #10 file sizes wrongly called sdreport
Found a different way to determine if bias adjustment was done.
I was calling sdreport too much and this should only be done
internally to optimize. I checked with VAST v.7... and it works
for determining if bias adjustment was or wasn't done.

Should still investigate if there are additional large objects that
are not needed.