feat: model NWFSC.Slope
The NWFSC.Slope survey is modelled very similar to the Triennial
The extrapolation grid is based on propInSlope02, which was the
largest footprint of those available for the propInSlope but
slightly smaller than the WCGBT Survey.
Only a few vessels were used, so overdispersion is not included by default.
Need to add an example.