docs: overdispersion
Use a template for overdispersion and combine all of the old information
and some new infomation gleaned from the VAST manual.
Realized that eta1 and delta1 led to the same result. It doesn't matter
what you name the parameters here, the end results in parameter_estimates.txt
is the same. I checked this by running two models and comparing the results
with the parameters named differently. I am not sure what VAST
actually does with the names.