issue #11 : use new function names for VAST and FishStatsUtils
* remove ability to do anything with alaska data
* remove ability to do anything with WCGOP data
* remove depth as a covariate
* add dataset
* change to using the most recent version of VAST and FishStatsUtils
* model estimates were checked with those for the 2019 sablefish assessment
that I had saved in my google drive. The results were matched down to 1.0e-5
in absolute value because JTT noted that some of the parameters are invertible.
There was a change in the Li_z parameter name that now includes eta, but the
estimate was the same. Gradients changed but AIC did not. I am not going to worry
about the gradient. The  model is much much faster now, running in 18 minutes
instead of an hour+.
* remove simulation capability -- I can look back if I want to add any of this
later but it isn't needed right now and the VAST model has a simulator

Need:
* maybe more diagnostic plots, I am not sure if JTT has all of the ones that
the council wants in plot_results()
* make some of our own plots
* summary of model run in .tex format to pull into the assessment document
* review the index vs index2 and move to the newest form of index2 and the
one argument that i have as FALSE for the extrapolation grid, fine_scale or
something or other.