feat: plot residuals using plot_results()
check_residuals = TRUE, for some reason now the saving of the .cpp
to a different folder and loading is allowing the residuals to be
checked.

wrapping in suppressWarnings removes the warning messages about the
projection, which Jim has already been told about and he wants to
leave them as is.