fix(VAST_spp): use common_name for species directory
This is under the assumption that people model their species every
year with the same name and then when you are done for the year
you would archive it with the year. For example, during the working
period a folder called "sablefish" would be the main directory and then
once you were done it would be archived as sablefish_2021.