feat: Saves model object if no standard errors are available
Before, if there were no standard errors the model would continue
instead of entering the early return. Errors/Warnings would arise
because in the plotting of the index the standard errors are needed.
We should not be doing diagnostics on models without inverted hessian
matrices; so, now `VAST_do()` will exit early if the model did not
converge rather than erroring out.