fix(Table_for_SS3): Finds appropriate columns in indexdata
Changes to VAST and FishStatsUtils that create the file
Index rather than Table_for_SS3 led to column name changes.
These legacy columns are easily recreated and saved similar to
the old table in Table_for_SS3.csv to allow downstream code to work.
This change was discussed in #39 with great changes made by
@chantelwetzel-NOAA but we decided to go with the simpler change here
rather than changing the column names of everything in VASTWestCoast
which would also lead to code changes that would be needed for
assessment purposes.