feat(use_anisotropy): Allows users to change TRUE
use_anisotropy = TRUE is the default in make_settings(),
which was previously used by VASTWestCoast, now
users can put FALSE in their get_settings() input list if
they want to chagne it. This turns off the estimation of the
H matrix that allows for spatial rotation, i.e., rotation
of the axis to be slanted rather than perpindicular to the coast.